- Added SystemInfo and ResourcesWindow for real-time resource usage monitoring - Renamed GenericTrack -> InMemoryTrack - A few other random cleanups

+++ src/square/ScrollableWindow.cpp
@@ -32,7 +32,7 @@ size_t ScrollableWindow::GetFirstVisible() {
 }
 
 size_t ScrollableWindow::GetLastVisible() {
-    int total = GetScrollAdapter().GetLineCount(this->GetContentWidth());
+    size_t total = GetScrollAdapter().GetLineCount(this->GetContentWidth());
     return min(scrollPosition + this->GetContentHeight(), total);
 }
